requirements.txt
slack_bolt~=1.20.1
python-dotenv~=1.0.1
gspread~=6.1.2
google-generativeai
google
